Controllerabstraction (#562)
* Controller abstraction

* Cartridge Provider

* Add cartridge rpc proxy

* Cleanup

* Use starknet-rs account factory

* Fix execute outside payload format (#530)

* Update starknet-rs

* Refactoring tests

* Update tests

* Fix execute_from_outside parsing

* Outside execution updates

* Cleanup impl

* Build wasm

* Fix session registration

* Fix session sign

* Update starknet-rs

* Pass app id

* Update starknet-rs

* fix: use cainome macro instead of bin generation

* Fix execute error after immediate deploy (#554)

* A few fixes

* Fix origin

* Fix controller fromStore

* Check session

* Controllerabstraction fixes (#559)

* Login/signup flow working

* Update slot session

* Update slot session

* clippy and fmt

* Remove old local storage session

* Fix execute outside payload

---------

+++ packages/keychain/src/utils/connection/probe.ts
@@ -2,8 +2,15 @@ import { ProbeReply, ResponseCodes } from "@cartridge/controller";
 import Controller from "utils/controller";
 
 export function probeFactory(setController: (controller: Controller) => void) {
-  return (controller: Controller, _origin: string) =>
+  return (origin: string) =>
     (rpcUrl?: string): Promise<ProbeReply> => {
+      const controller = Controller.fromStore(origin);
+      if (!controller) {
+        return Promise.reject({
+          code: ResponseCodes.NOT_CONNECTED,
+        });
+      }
+
       if (rpcUrl && rpcUrl !== controller.rpcUrl) {
         controller.delete();
         setController(undefined);
@@ -12,6 +19,7 @@ export function probeFactory(setController: (controller: Controller) => void) {
         });
       }
 
+      setController(controller);
       return Promise.resolve({
         code: ResponseCodes.SUCCESS,
         address: controller.address,
